backend/gladio_backend/requirements.txt
Django
Django-ninja
mysqlclient
django-cors-headers
firebase-admin